价格+粉丝id 类型修改

+++ database/migrations/2018_11_12_041602_create_mall_goods_table.php
@@ -17,11 +17,11 @@ class CreateMallGoodsTable extends Migration
             $table->increments('id');
             $table->string('name',100)->comment('名称');
             $table->longText('content')->comment('详细描述');
-            $table->tinyInteger('total')->comment('商品数量');
-            $table->tinyInteger('limit')->comment('用户购买上限');
-            $table->tinyInteger('price')->comment('价格');
-            $table->tinyInteger('discount')->comment('优惠价');
-            $table->tinyInteger('monthly_sales')->comment('月销量');
+            $table->integer('total')->comment('商品数量');
+            $table->integer('limit')->comment('用户购买上限');
+            $table->decimal('price')->comment('价格');
+            $table->decimal('discount')->comment('优惠价');
+            $table->integer('monthly_sales')->comment('月销量');
             $table->tinyInteger('is_up')->default(1)->comment('是否上架,1表是，0表否');
             $table->dateTime('sratr_date')->comment('生效日期');
             $table->dateTime('end_date')->comment('失效日期');
