+++ database/migrations/2019_01_24_110455_create_tickets_table.php
@@ -19,12 +19,12 @@ class CreateTicketsTable extends Migration
             $table->string('cover', 512)->default('')->comment('封面');
             $table->longText('content')->comment('详细描述');
             $table->string('remark',300)->commend('备注');
-            $table->tinyInteger('total')->comment('每日数量');
-            $table->tinyInteger('limit')->comment('用户每日购买上限');
-            $table->tinyInteger('daily_inventory')->comment('每日库存');
-            $table->tinyInteger('price')->comment('价格');
+            $table->integer('total')->comment('每日数量');
+            $table->integer('limit')->comment('用户每日购买上限');
+            $table->integer('daily_inventory')->comment('每日库存');
+            $table->decimal('price')->comment('价格');
             $table->tinyInteger('is_up')->default(1)->comment('是否上架,1表是，0表否');
-            $table->tinyInteger('from_now')->comment('限制天数');
+            $table->integer('from_now')->comment('限制天数');
             $table->timestamps();
         });
     }
